removed old blanket from plasma
