+++ paramak/__init__.py
@@ -23,9 +23,6 @@ from .parametric_components.blanket_constant_thickness_arc_v import (
 from .parametric_components.blanket_constant_thickness_fp import (
     BlanketConstantThicknessFP,
 )
-from .parametric_components.blanket_constant_thickness_plasma import (
-    BlanketConstantThicknessPlasma,
-)
 
 from .parametric_components.divertor_block import DivertorBlock
