+++ paramak/parametric_components/blanket_constant_thickness_fp.py
@@ -1,28 +1,56 @@
-import scipy
 import math
 
 import numpy as np
+import scipy
+import sympy as sp
 
 from paramak import RotateMixedShape
 
 
 class BlanketConstantThicknessFP(RotateMixedShape):
-    """An outboard blanket volume that follows the curvature of the plasma
-    with a fixed offset from the plasma and a constant blanket thickness.
-    The blanket volume has start and stop angles that allow the blanket 
-    coverage to be increased or decreased.
-
-    :param plasma: a plasma object
-    :type plasma: paramak Plasma object
+    """
     :param thickness:  the thickness of the blanket (cm)
     :type thickness: float
-    :param stop_angle: the angle in degrees to stop the blanket, measured anti clockwise from 3 o'clock
+    :param stop_angle: the angle in degrees to stop the blanket, measured anti
+     clockwise from 3 o'clock
     :type stop_angle: float
-    :param start_angle: the angle in degrees to start the blanket, measured anti clockwise from 3 o'clock
+    :param start_angle: the angle in degrees to start the blanket, measured
+     anti clockwise from 3 o'clock
     :type start_angle: float
+    :param minor_radius: the minor radius of the plasma (cm)
+    :type minor_radius: float
+    :param major_radius: the major radius of the plasma (cm)
+    :type major_radius: float
+    :param triangularity: the triangularity of the plasma
+    :type triangularity: float
+    :param elongation: the elongation of the plasma
+    :type elongation: float
+    :param vertical_displacement: the vertical_displacement of the plasma (cm)
+    :type vertical_displacement: float
     :param offset_from_plasma: the distance bettwen the plasma and the blanket (cm)
     :type offset_from_plasma: float
-
+    :param num_points: number of points that will describe the shape
+    :type num_points: int
+    :param name: The legend name used when exporting a html graph of the shape
+    :type name: str
+    :param color: the color to use when exporting as html graphs or png images
+    :type color: Red, Green, Blue, [Alpha] values. RGB and RGBA are sequences
+     of, 3 or 4 floats respectively each in the range 0-1
+    :param material_tag: The material name to use when exporting the
+     neutronics description
+    :type material_tag: str
+    :param stp_filename: the filename used when saving stp files as part of a
+     reactor
+    :type stp_filename: str
+    :param azimuth_placement_angle: the angle or angles to use when rotating
+     the shape on the azimuthal axis
+    :type azimuth_placement_angle: float or iterable of floats
+    :param rotation_angle: The rotation_angle to use when revoling the solid
+     (degrees)
+    :type rotation_angle: float
+    :param cut: An optional cadquery object to perform a boolean cut with this
+     object
+    :type cut: cadquery object
 
     :return: a shape object that has generic functionality
     :rtype: paramak shape object
@@ -30,11 +58,17 @@ class BlanketConstantThicknessFP(RotateMixedShape):
 
     def __init__(
         self,
-        plasma,
         thickness,
-        stop_angle,
         start_angle,
-        offset_from_plasma,
+        stop_angle,
+        plasma=None,
+        minor_radius=150,
+        major_radius=450,
+        triangularity=0.55,
+        elongation=2.0,
+        vertical_displacement=0,
+        offset_from_plasma=0,
+        num_points=50,
         workplane="XZ",
         points=None,
         stp_filename=None,
@@ -45,7 +79,7 @@ class BlanketConstantThicknessFP(RotateMixedShape):
         name=None,
         material_tag=None,
         cut=None,
-    ):
+            ):
 
         super().__init__(
             points,
@@ -60,14 +94,24 @@ class BlanketConstantThicknessFP(RotateMixedShape):
             cut,
         )
 
-        self.major_radius = plasma.major_radius
-        self.minor_radius = plasma.minor_radius
-        self.triangularity = plasma.triangularity
-        self.elongation = plasma.elongation
         self.thickness = thickness
-        self.stop_angle = stop_angle
         self.start_angle = start_angle
-        self.offset_from_plasma = offset_from_plasma
+        self.stop_angle = stop_angle
+        self.plasma = plasma
+        self.vertical_displacement = vertical_displacement
+        if plasma is None:
+            self.minor_radius = minor_radius
+            self.major_radius = major_radius
+            self.triangularity = triangularity
+            self.elongation = elongation
+            self.offset_from_plasma = 0
+        else:  # if plasma object is given, use its parameters
+            self.minor_radius = plasma.minor_radius
+            self.major_radius = plasma.major_radius
+            self.triangularity = plasma.triangularity
+            self.elongation = plasma.elongation
+            self.offset_from_plasma = offset_from_plasma
+        self.num_points = num_points
         self.points = points
 
     @property
@@ -80,36 +124,12 @@ class BlanketConstantThicknessFP(RotateMixedShape):
         self._points = points
 
     @property
-    def major_radius(self):
-        return self._major_radius
-
-    @major_radius.setter
-    def major_radius(self, major_radius):
-        self._major_radius = major_radius
-
-    @property
     def minor_radius(self):
         return self._minor_radius
 
     @minor_radius.setter
     def minor_radius(self, minor_radius):
         self._minor_radius = minor_radius
-
-    @property
-    def triangularity(self):
-        return self._triangularity
-
-    @triangularity.setter
-    def triangularity(self, triangularity):
-        self._triangularity = triangularity
-
-    @property
-    def elongation(self):
-        return self._elongation
-
-    @elongation.setter
-    def elongation(self, elongation):
-        self._elongation = elongation
 
     @property
     def thickness(self):
@@ -119,88 +139,102 @@ class BlanketConstantThicknessFP(RotateMixedShape):
     def thickness(self, thickness):
         self._thickness = thickness
 
-    @property
-    def stop_angle(self):
-        return self._stop_angle
-
-    @stop_angle.setter
-    def stop_angle(self, stop_angle):
-        self._stop_angle = stop_angle
-
-    @property
-    def start_angle(self):
-        return self._start_angle
-
-    @start_angle.setter
-    def start_angle(self, start_angle):
-        self._start_angle = start_angle
-
-    @property
-    def offset_from_plasma(self):
-        return self._offset_from_plasma
-
-    @offset_from_plasma.setter
-    def offset_from_plasma(self, offset_from_plasma):
-        self._offset_from_plasma = offset_from_plasma
-
-    @property
-    def inner_limit(self):
-        return self._inner_limit
-
-    @inner_limit.setter
-    def inner_limit(self, inner_limit):
-        self._inner_limit = inner_limit
-
     def find_points(self):
-        """Finds the XZ points and connection types (straight and spline) that
-        describe the 2D profile of the blanket shape."""
+        conversion_factor = 2*np.pi/360
 
-        # This section finds the points on the front face of the divertor
-        radius_of_front_curve = (
-            0.5
-            * scipy.sqrt(
-                (
-                    self.minor_radius ** 2
-                    * ((self.triangularity + 1) ** 2 + self.elongation ** 2) ** 2
-                )
-                / ((self.triangularity + 1) ** 2)
-            )
-            + self.offset_from_plasma
-        )
+        def R(theta, pkg=np):
+            return self.major_radius + self.minor_radius*pkg.cos(
+                theta + self.triangularity*pkg.sin(theta))
 
-        x_position_for_outside_arc = (
-            2 * self.major_radius * (1 + self.triangularity)
-            - self.minor_radius * (self.triangularity ** 2 + self.elongation ** 2 - 1)
-        ) / (2 * (1 + self.triangularity))
+        def Z(theta, pkg=np):
+            return self.elongation*self.minor_radius*pkg.sin(theta) + \
+                self.vertical_displacement
+        # create array of angles theta
+        thetas = np.linspace(
+                    self.start_angle*conversion_factor,
+                    self.stop_angle*conversion_factor,
+                    num=self.num_points,
+                    endpoint=True)
 
-        angles = np.linspace(
-            math.radians(self.start_angle), math.radians(self.stop_angle), 100
-        )
+        # create sympy objects and derivatives
+        theta_sp = sp.Symbol("theta")
 
-        xs = -(radius_of_front_curve * scipy.cos(angles) - x_position_for_outside_arc)
-        zs = radius_of_front_curve * scipy.sin(angles)
+        R_sp = R(theta_sp, pkg=sp)
+        Z_sp = Z(theta_sp, pkg=sp)
+
+        R_derivative = sp.diff(R_sp, theta_sp)
+        Z_derivative = sp.diff(Z_sp, theta_sp)
+        # create inner points
+        if self.plasma is None:
+            # if no plasma object is given simply use the equation
+            inner_points_R = R(thetas)
+            inner_points_Z = Z(thetas)
+
+            points = [
+                [inner_points_R[i], inner_points_Z[i], 'spline']
+                for i in range(len(thetas))
+                ]
+        else:
+            # if a plasma is given
+            points = self.create_offset_points(
+                thetas, R, Z,
+                R_derivative, Z_derivative,
+                self.offset_from_plasma)
+        points[-1][2] = 'straight'
+
+        # compute outer points
+        outer_points = self.create_offset_points(
+            thetas, R, Z,
+            R_derivative, Z_derivative,
+            self.thickness + self.offset_from_plasma, flip=True)
+        points = points + outer_points
+        points[-2][2] = 'straight'
+        self.points = points
+
+    def create_offset_points(
+            self, thetas, R_fun, Z_fun, R_derivative, Z_derivative, offset,
+            flip=False):
+        """generates a list of points following parametric equations with an
+            offset.
+
+        Args:
+            thetas (list): list of angles (radiants)
+            R_fun (callable): parametric function for R coordinate (cm)
+            Z_fun ([type]): parametric function for Z coordinate (cm)
+            R_derivative (sympy.Mul): derivative of R over theta (cm/rad)
+            Z_derivative (sympy.Mul): derivative of Z over theta (cm/rad)
+            offset (float): offset value (cm). offset=0 will follow the
+                parametric equations.
+            flip (bool, optional): If True thetas will be iterated from the
+                end. Defaults to False.
+
+        Returns:
+            list: list of points
+                [[R1, Z1, connection1], [R2, Z2, connection2], ...]
+        """
 
         points = []
-        for x, z, in zip(xs, zs):
-            points.append([x, z, "spline"])
+        list_of_thetas = thetas
+        if flip:
+            list_of_thetas = np.flip(thetas)
 
-        points[-1][2] = "straight"
+        for theta in list_of_thetas:
+            # get local value of derivatives
+            val_R_derivative = float(R_derivative.subs('theta', theta))
+            val_Z_derivative = float(Z_derivative.subs('theta', theta))
 
-        # This section finds the points on the rear face of the divertor
-        radius_of_back_curve = radius_of_front_curve + self.thickness
+            # get normal vector components
+            nx = val_Z_derivative
+            ny = -val_R_derivative
 
-        # stop angle is converted from degrees to radians
-        angles = np.linspace(
-            math.radians(self.stop_angle), math.radians(self.start_angle), 100
-        )
+            # normalise normal vector
+            normal_vector_norm = (nx**2 + ny**2)**0.5
+            nx /= normal_vector_norm
+            ny /= normal_vector_norm
 
-        xs = -(radius_of_back_curve * scipy.cos(angles) - x_position_for_outside_arc)
-        zs = radius_of_back_curve * scipy.sin(angles)
+            # calculate outer points
+            val_R_outer = R_fun(theta) + offset*nx
+            val_Z_outer = Z_fun(theta) + offset*ny
 
-        for x, z, in zip(xs, zs):
-            points.append([x, z, "spline"])
-
-        # changes the last point to a straght conenctor
-        points[-1][2] = "straight"
-
-        self.points = points
+            points.append([float(val_R_outer), float(val_Z_outer), 'spline'])
+        return points
