+++ tests/test_ParametricComponents.py
@@ -76,8 +76,8 @@ class test_BlanketConstantThickness(unittest.TestCase):
         assert test_shape.volume > 1000
 
 
-class test_BlanketConstantThicknessPlasma(unittest.TestCase):
-    def test_BlanketConstantThicknessPlasma_creation_plasma(self):
+class test_BlanketConstantThicknessFP(unittest.TestCase):
+    def test_BlanketConstantThicknessFP_creation_plasma(self):
         """creates blanket from parametric shape and checks a solid is created"""
         plasma = paramak.Plasma(
             major_radius=300,
@@ -85,7 +85,7 @@ class test_BlanketConstantThicknessPlasma(unittest.TestCase):
             triangularity=0.5,
             elongation=2,
         )
-        test_shape = paramak.BlanketConstantThicknessPlasma(
+        test_shape = paramak.BlanketConstantThicknessFP(
             plasma=plasma,
             thickness=200,
             stop_angle=90,
@@ -96,10 +96,10 @@ class test_BlanketConstantThicknessPlasma(unittest.TestCase):
         assert test_shape.solid is not None
         assert test_shape.volume > 1000
 
-    def test_BlanketConstantThicknessPlasma_creation_noplasma(self):
+    def test_BlanketConstantThicknessFP_creation_noplasma(self):
         """creates blanket from parametric shape and checks a solid is created"""
 
-        test_shape = paramak.BlanketConstantThicknessPlasma(
+        test_shape = paramak.BlanketConstantThicknessFP(
             major_radius=300,
             minor_radius=50,
             triangularity=0.5,
